Relocate metrics tracker into plugin reporting
