+++ src/plugin/src/identifier-case/local-plan.js
@@ -3,7 +3,7 @@ import { buildRenameKey } from "./plan-state.js";
 import { asArray, isNonEmptyArray } from "../../../shared/array-utils.js";
 import { toPosixPath } from "../../../shared/path-utils.js";
 import { resolveProjectPathInfo } from "../project-index/path-info.js";
-import { createMetricsTracker } from "../../../shared/metrics-utils.js";
+import { createMetricsTracker } from "../reporting/metrics.js";
 import {
     isNonEmptyString,
     getNonEmptyString,
