+++ src/plugin/src/project-index/metrics.js
@@ -1,4 +1,4 @@
-import { createMetricsTracker } from "../../../shared/metrics-utils.js";
+import { createMetricsTracker } from "../reporting/metrics.js";
 
 const PROJECT_INDEX_METRICS_CATEGORY = "project-index";
 const REQUIRED_METRIC_METHODS = [
