+++ src/plugin/src/reporting/index.js
@@ -1,1 +1,1 @@
-export * from "../../../shared/metrics-utils.js";
+export * from "./metrics.js";
