+++ src/plugin/tests/reporting-metrics.test.js
@@ -1,7 +1,7 @@
 import assert from "node:assert/strict";
 import test from "node:test";
 
-import { createMetricsTracker } from "../../shared/metrics-utils.js";
+import { createMetricsTracker } from "../src/reporting/metrics.js";
 
 test("snapshot exposes accumulated metrics as plain objects", () => {
     const tracker = createMetricsTracker({ category: "demo" });
